+++ src/client/search.rs
@@ -6,6 +6,7 @@ use crate::{
 
 use chrono::{DateTime, FixedOffset};
 use failure::Fail;
+use log::debug;
 use mime;
 use reqwest::{Response, StatusCode};
 use serde::{self, Serialize, Deserialize};
@@ -201,14 +202,18 @@ pub fn search_documents(authorized_client: &AuthorizedClient, search: Search) ->
 
     let internal_search = internal::Search::from_search(search);
 
-    let mut response: Response = authorized_client
+    let request = authorized_client
         .http_client
         .post(&url)
         .bearer_auth(&authorized_client.token.access_token)
-        .json(&internal_search)
+        .json(&internal_search);
+    debug!("Request: '{:#?}'", request);
+
+    let mut response: Response = request
         .send()
         .map_err(|e| e.context(ErrorKind::HttpRequestFailed))?
         .general_err_handler(StatusCode::OK)?;
+    debug!("Response: '{:#?}'", response);
 
     let result = response.json().map_err(|e| e.context(ErrorKind::FailedToProcessHttpResponse(response.status(), "reading body".to_string())))?;